cleaned up the imports in the main workflow scripts. Minor bug fixes in condor_SUEP_WS.py and the inference. Added ROC curve to the plotter.
